requirements.txt
altgraph==0.15
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
itchat==1.3.10
macholib==1.9
olefile==0.44
pefile==2017.11.5
Pillow==4.3.0
PyInstaller==3.3.1
pypng==0.0.18
PyQRCode==1.2.1
requests==2.18.4
urllib3==1.22
